soal_2/client_2/requirements.txt
djangorestframework
djangorestframework-simplejwt==5.3.1
python-dotenv